requirements.txt
facebook-sdk
Pillow==4.2.1